arguments types improvement
git-svn-id: https://svn.code.sf.net/p/gopherus/code/trunk@54 767d67df-c2ba-418b-950d-eab06a56d38d

+++ net/net-lin.c
@@ -41,7 +41,7 @@ int net_init(void) {
 }
 
 
-struct net_tcpsocket *net_connect(unsigned long ipaddr, int port) {
+struct net_tcpsocket *net_connect(unsigned long ipaddr, unsigned short port) {
   struct netwrap *s;
   struct sockaddr_in remote;
   struct net_tcpsocket *result;
@@ -76,7 +76,7 @@ struct net_tcpsocket *net_connect(unsigned long ipaddr, int port) {
 
 /* Sends data on socket 'socket'.
    Returns the number of bytes sent on success, and <0 otherwise. The error code can be translated into a human error message via libtcp_strerr(). */
-int net_send(struct net_tcpsocket *socket, char *line, int len) {
+int net_send(struct net_tcpsocket *socket, char *line, long len) {
   int res;
   res = send(((struct netwrap *)(socket->sock))->fd, line, len, 0);
   return(res);
@@ -85,7 +85,7 @@ int net_send(struct net_tcpsocket *socket, char *line, int len) {
 
 /* Reads data from socket 'sock' and write it into buffer 'buff', until end of connection. Will fall into error if the amount of data is bigger than 'maxlen' bytes.
 Returns the amount of data read (in bytes) on success, or a negative value otherwise. The error code can be translated into a human error message via libtcp_strerr(). */
-int net_recv(struct net_tcpsocket *socket, char *buff, int maxlen) {
+int net_recv(struct net_tcpsocket *socket, char *buff, long maxlen) {
   int res;
   int realsocket;
   fd_set rfds;
